fix: support initial query has array string type

+++ utils/index.ts
@@ -9,14 +9,17 @@ export function isEmptyObj(obj: Record<string, unknown>): boolean {
 export function buildQuerySting(
   key: string,
   value?: string,
-  initial: Record<string, string> = {}
+  initial: Record<string, string | string[]> = {}
 ): string {
-  const query = new URLSearchParams(initial)
-  if (value) {
-    query.set(key, value)
-  } else {
-    query.delete(key)
+  const query = new URLSearchParams()
+  if (initial) {
+    for (const [key, value] of Object.entries(initial)) {
+      if (typeof value === 'string') query.append(key, value)
+      if (Array.isArray(value)) value.forEach((v) => query.append(key, v))
+    }
   }
+  if (value) query.set(key, value)
+  if (!value || value === '#') query.delete(key)
   const queryStr = query.toString()
   return queryStr ? `?${queryStr}` : ''
 }
